Basic Controls
This code did not delete the old code, it simply commented out what was not being tested. This code should provide basic control of adding a dig_pot value. The old code was also evaluated for bugs and some big ones were found. If the new control works the changes to the old code may work as well.